Small fix for XMPP mock server
In the last commit, the XMPP mock server allows rules like the
following:

{
  ...
  '<iq type="set">...</iq>':
  {
    '':
    '<iq type="result">...</iq>'.

    '<iq type="get">...</iq>':
    '<iq type="result">...</iq>'
  }
}

which changes the set of rules in reaction to a specific query. The
empty-string rule ('') is the immediate reply to the query, but was
accidently added to the rule table, which leads to problems. This
is fixed now.