Premilinary support for proxying media server queries
Works by redirecting the client to the real media server resource
and handling the authentication. Due to a currently missing way
of discovering a media server, this only works for channels in
the HTTP API's home domain for now. Also, the media server still
needs to be fixed to accept anonymous queries, so the feature
works only authenticated sessions.