Remove channel server discovery, add channelDomain option
It turns out that even if communicating with a remote channel server,
the buddycloud protocol expects the request to go to the user's
home channel server at all times. Therefore, I removed the discovery
code and just added an option to specify your home channel server's
component domain. This also fixes the bug that remote servers could
not be accessed.